requirements.txt
langchain==0.1.11
docarray==0.40.0
tiktoken==0.6.0
langchain_openai==0.0.8
python-dotenv==1.0.1
pypdf==4.1.0
chromadb==0.4.24
Flask==3.0.2
python-docx==1.1.0
docx2txt==0.8